v2.4.5: fixed pitch and volume always going to their default values

+++ core/src/main/java/dev/endoy/djp/storage/AbstractStorageManager.java
@@ -88,9 +88,9 @@ public abstract class AbstractStorageManager
 
     public abstract void setSlotGroup( UUID uuid, String slotGroup );
 
-    public abstract void setJoinSound( UUID uuid, String sound, float volume, float pitch );
+    public abstract void setJoinSound( UUID uuid, String sound, Integer volume, Integer pitch );
 
-    public abstract void setLeaveSound( UUID uuid, String sound, float volume, float pitch );
+    public abstract void setLeaveSound( UUID uuid, String sound, Integer volume, Integer pitch );
 
     public abstract void toggleSound( UUID uuid, boolean toggled );
